🔨 REFACTOR: SettingsPage
